➖ Deleted `git.rs` file
➕ Added `gitmoji.rs` file

♻️ Refactored `mod.rs` and `utils.rs` in the `git` module

✨ Added `replace` function to `gitmoji.rs` for replacing gitmoji tags in commit messages

:hammer_and_wrench: Updated `utils.rs` to include `get_branch_name` function
