+++ src/git/mod.rs
@@ -1,0 +1,7 @@
+mod commands;
+pub mod commits;
+pub mod gitmoji;
+mod utils;
+
+pub use self::commands::{cached_diff, checkout, squash, stash};
+pub use self::utils::get_branch_name;
